requirements.txt
py-healthcheck
Flask
psutil
mysql-connector-python
schedule
python-dotenv 